# app/models.py
# generated by fastapi-codegen:
#   filename:  openapi/schema.yml
#   timestamp: 2022-11-30T11:50:22+00:00

from __future__ import annotations

from typing import List, Optional

from pydantic import BaseModel, Field


class Pet(BaseModel):
    id: int
    name: str
    tag: Optional[str] = None


class Pets(BaseModel):
    __root__: List[Pet] = Field(..., description='list of pet')


class Error(BaseModel):
    code: int
    message: str